Initialize Database in splash screen
This creates a new activity called SplashActivity.
SplashActivity waits until all the initialization for the Database is done
before starting MainActivity.
This is done to fix issues where fragments where calling getDeviceUser() before
the user was initialized properly.

This also adds a parameter to Database.initializeInstance() which is a callback
that is called when all the initialization is finished.